ai-platform-react-python/backend/requirements.txt
# Flask 核心
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3

# 資料庫
pymongo==4.6.0
redis==5.0.1

# Google Cloud AI Platform
google-generativeai==0.3.2
google-cloud-aiplatform==1.40.0
google-cloud-storage==2.14.0
numpy==1.24.3

# 文件處理
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3

# 數據驗證
pydantic==2.5.0

# 密碼處理
passlib==1.7.4
bcrypt==4.1.2

# 工具
python-dotenv==1.0.0
werkzeug==3.0.1
langdetect==1.0.9

# 日誌和監控
python-json-logger==2.0.7